Fixed problem using reshape, made an example script
Had to switch instances of arr.reshape(size) to arr.reshape(size, order='F') in order to get behaviour consistent with the vectorization operator, Fortran-like indexing. The default behaviour was C-like indexing which is different from MATLAB, and caused an incorrect solution to be returned.

Example use shown in Python\example_usage.py.